[FIX] Split large databases composite tool
Extracted each database action category into its own module within
src/tools/composite/databases/ to reduce complexity and improve
maintainability.

- Created src/tools/composite/databases/types.ts for shared interfaces.
- Created src/tools/composite/databases/helpers.ts for shared logic and caching.
- Created src/tools/composite/databases/core.ts for database creation and metadata.
- Created src/tools/composite/databases/query.ts for database querying.
- Created src/tools/composite/databases/pages.ts for item CRUD operations.
- Created src/tools/composite/databases/data-sources.ts for data source management.
- Refactored src/tools/composite/databases.ts as a clean entry point delegating to the modules.

Verified all 53 database tests and the full suite of 734 tests pass.
Ran bun run check:fix to ensure linting and type-safety.
